+++ src/ralph/models/edx/__init__.py
@@ -10,4 +10,5 @@ from .enrollment import (
     UIEdxCourseEnrollmentUpgradeClicked,
 )
 from .navigational import UIPageClose, UISeqGoto, UISeqNext, UISeqPrev
+from .open_response_assessment import ORASaveSubmission
 from .server import ServerEvent